beautify the code(added pokemon.js file) with addition of new pokemon and correction of console.log in app.listen with backticks
